Implemented point lights & directional lights in classes & shader
Now the shader supports using multiple point lights and multiple directional lights (for now 32 maximum of each, though value can be easily changed). Need to add spotlights, as well as a LightManager that can easily be added to a Renderer and sets up all the lights added to it automatically and easily.